Convex hull

Computes the convex hull of a set of points.

Usage

cxhull(points, triangulate = FALSE)

Arguments

points

numeric matrix, one point per row

triangulate

logical, whether to triangulate the convex hull

Value

A list providing a lot of information about the convex hull. See the README file for details.

Examples

library(cxhull)
points <- rbind(
 c(0.5,0.5,0.5),
 c(0,0,0),
 c(0,0,1),
 c(0,1,0),
 c(0,1,1),
 c(1,0,0),
 c(1,0,1),
 c(1,1,0),
 c(1,1,1)
)
cxhull(points)
#> $vertices
#> $vertices$`2`
#> $vertices$`2`$id
#> [1] 2
#> 
#> $vertices$`2`$point
#> [1] 0 0 0
#> 
#> $vertices$`2`$neighvertices
#> [1] 3 4 6
#> 
#> $vertices$`2`$neighridges
#> [1] 1 2 5
#> 
#> $vertices$`2`$neighfacets
#> [1] 1 2 4
#> 
#> 
#> $vertices$`3`
#> $vertices$`3`$id
#> [1] 3
#> 
#> $vertices$`3`$point
#> [1] 0 0 1
#> 
#> $vertices$`3`$neighvertices
#> [1] 2 5 7
#> 
#> $vertices$`3`$neighridges
#> [1]  5  6 10
#> 
#> $vertices$`3`$neighfacets
#> [1] 2 4 6
#> 
#> 
#> $vertices$`4`
#> $vertices$`4`$id
#> [1] 4
#> 
#> $vertices$`4`$point
#> [1] 0 1 0
#> 
#> $vertices$`4`$neighvertices
#> [1] 2 5 8
#> 
#> $vertices$`4`$neighridges
#> [1]  1  3 11
#> 
#> $vertices$`4`$neighfacets
#> [1] 1 4 5
#> 
#> 
#> $vertices$`5`
#> $vertices$`5`$id
#> [1] 5
#> 
#> $vertices$`5`$point
#> [1] 0 1 1
#> 
#> $vertices$`5`$neighvertices
#> [1] 3 4 9
#> 
#> $vertices$`5`$neighridges
#> [1] 10 11 12
#> 
#> $vertices$`5`$neighfacets
#> [1] 4 5 6
#> 
#> 
#> $vertices$`6`
#> $vertices$`6`$id
#> [1] 6
#> 
#> $vertices$`6`$point
#> [1] 1 0 0
#> 
#> $vertices$`6`$neighvertices
#> [1] 2 7 8
#> 
#> $vertices$`6`$neighridges
#> [1] 2 4 7
#> 
#> $vertices$`6`$neighfacets
#> [1] 1 2 3
#> 
#> 
#> $vertices$`7`
#> $vertices$`7`$id
#> [1] 7
#> 
#> $vertices$`7`$point
#> [1] 1 0 1
#> 
#> $vertices$`7`$neighvertices
#> [1] 3 6 9
#> 
#> $vertices$`7`$neighridges
#> [1] 6 7 9
#> 
#> $vertices$`7`$neighfacets
#> [1] 2 3 6
#> 
#> 
#> $vertices$`8`
#> $vertices$`8`$id
#> [1] 8
#> 
#> $vertices$`8`$point
#> [1] 1 1 0
#> 
#> $vertices$`8`$neighvertices
#> [1] 4 6 9
#> 
#> $vertices$`8`$neighridges
#> [1] 3 4 8
#> 
#> $vertices$`8`$neighfacets
#> [1] 1 3 5
#> 
#> 
#> $vertices$`9`
#> $vertices$`9`$id
#> [1] 9
#> 
#> $vertices$`9`$point
#> [1] 1 1 1
#> 
#> $vertices$`9`$neighvertices
#> [1] 5 7 8
#> 
#> $vertices$`9`$neighridges
#> [1]  8  9 12
#> 
#> $vertices$`9`$neighfacets
#> [1] 3 5 6
#> 
#> 
#> 
#> $edges
#>       [,1] [,2]
#>  [1,]    2    3
#>  [2,]    2    4
#>  [3,]    2    6
#>  [4,]    3    5
#>  [5,]    3    7
#>  [6,]    4    5
#>  [7,]    4    8
#>  [8,]    5    9
#>  [9,]    6    7
#> [10,]    6    8
#> [11,]    7    9
#> [12,]    8    9
#> 
#> $ridges
#> $ridges[[1]]
#> $ridges[[1]]$id
#> [1] 1
#> 
#> $ridges[[1]]$ridgeOf
#> [1] 1 4
#> 
#> $ridges[[1]]$vertices
#> [1] 2 4
#> 
#> 
#> $ridges[[2]]
#> $ridges[[2]]$id
#> [1] 2
#> 
#> $ridges[[2]]$ridgeOf
#> [1] 1 2
#> 
#> $ridges[[2]]$vertices
#> [1] 2 6
#> 
#> 
#> $ridges[[3]]
#> $ridges[[3]]$id
#> [1] 3
#> 
#> $ridges[[3]]$ridgeOf
#> [1] 1 5
#> 
#> $ridges[[3]]$vertices
#> [1] 4 8
#> 
#> 
#> $ridges[[4]]
#> $ridges[[4]]$id
#> [1] 4
#> 
#> $ridges[[4]]$ridgeOf
#> [1] 1 3
#> 
#> $ridges[[4]]$vertices
#> [1] 6 8
#> 
#> 
#> $ridges[[5]]
#> $ridges[[5]]$id
#> [1] 5
#> 
#> $ridges[[5]]$ridgeOf
#> [1] 2 4
#> 
#> $ridges[[5]]$vertices
#> [1] 2 3
#> 
#> 
#> $ridges[[6]]
#> $ridges[[6]]$id
#> [1] 6
#> 
#> $ridges[[6]]$ridgeOf
#> [1] 2 6
#> 
#> $ridges[[6]]$vertices
#> [1] 3 7
#> 
#> 
#> $ridges[[7]]
#> $ridges[[7]]$id
#> [1] 7
#> 
#> $ridges[[7]]$ridgeOf
#> [1] 2 3
#> 
#> $ridges[[7]]$vertices
#> [1] 6 7
#> 
#> 
#> $ridges[[8]]
#> $ridges[[8]]$id
#> [1] 8
#> 
#> $ridges[[8]]$ridgeOf
#> [1] 3 5
#> 
#> $ridges[[8]]$vertices
#> [1] 8 9
#> 
#> 
#> $ridges[[9]]
#> $ridges[[9]]$id
#> [1] 9
#> 
#> $ridges[[9]]$ridgeOf
#> [1] 3 6
#> 
#> $ridges[[9]]$vertices
#> [1] 7 9
#> 
#> 
#> $ridges[[10]]
#> $ridges[[10]]$id
#> [1] 10
#> 
#> $ridges[[10]]$ridgeOf
#> [1] 4 6
#> 
#> $ridges[[10]]$vertices
#> [1] 3 5
#> 
#> 
#> $ridges[[11]]
#> $ridges[[11]]$id
#> [1] 11
#> 
#> $ridges[[11]]$ridgeOf
#> [1] 4 5
#> 
#> $ridges[[11]]$vertices
#> [1] 4 5
#> 
#> 
#> $ridges[[12]]
#> $ridges[[12]]$id
#> [1] 12
#> 
#> $ridges[[12]]$ridgeOf
#> [1] 5 6
#> 
#> $ridges[[12]]$vertices
#> [1] 5 9
#> 
#> 
#> 
#> $facets
#> $facets[[1]]
#> $facets[[1]]$vertices
#> [1] 8 4 6 2
#> 
#> $facets[[1]]$edges
#>      [,1] [,2]
#> [1,]    2    4
#> [2,]    2    6
#> [3,]    4    8
#> [4,]    6    8
#> 
#> $facets[[1]]$ridges
#> [1] 1 2 3 4
#> 
#> $facets[[1]]$neighbors
#> [1] 2 3 4 5
#> 
#> $facets[[1]]$volume
#> [1] 1
#> 
#> $facets[[1]]$center
#> [1] 0.5 0.5 0.0
#> 
#> $facets[[1]]$normal
#> [1]  0  0 -1
#> 
#> $facets[[1]]$offset
#> [1] 0
#> 
#> $facets[[1]]$family
#> [1] NA
#> 
#> $facets[[1]]$orientation
#> [1] 1
#> 
#> 
#> $facets[[2]]
#> $facets[[2]]$vertices
#> [1] 7 3 6 2
#> 
#> $facets[[2]]$edges
#>      [,1] [,2]
#> [1,]    2    3
#> [2,]    2    6
#> [3,]    3    7
#> [4,]    6    7
#> 
#> $facets[[2]]$ridges
#> [1] 2 5 6 7
#> 
#> $facets[[2]]$neighbors
#> [1] 1 3 4 6
#> 
#> $facets[[2]]$volume
#> [1] 1
#> 
#> $facets[[2]]$center
#> [1] 0.5 0.0 0.5
#> 
#> $facets[[2]]$normal
#> [1]  0 -1  0
#> 
#> $facets[[2]]$offset
#> [1] 0
#> 
#> $facets[[2]]$family
#> [1] NA
#> 
#> $facets[[2]]$orientation
#> [1] -1
#> 
#> 
#> $facets[[3]]
#> $facets[[3]]$vertices
#> [1] 7 8 9 6
#> 
#> $facets[[3]]$edges
#>      [,1] [,2]
#> [1,]    6    7
#> [2,]    6    8
#> [3,]    7    9
#> [4,]    8    9
#> 
#> $facets[[3]]$ridges
#> [1] 4 7 8 9
#> 
#> $facets[[3]]$neighbors
#> [1] 1 2 5 6
#> 
#> $facets[[3]]$volume
#> [1] 1
#> 
#> $facets[[3]]$center
#> [1] 1.0 0.5 0.5
#> 
#> $facets[[3]]$normal
#> [1] 1 0 0
#> 
#> $facets[[3]]$offset
#> [1] -1
#> 
#> $facets[[3]]$family
#> [1] NA
#> 
#> $facets[[3]]$orientation
#> [1] 1
#> 
#> 
#> $facets[[4]]
#> $facets[[4]]$vertices
#> [1] 5 3 4 2
#> 
#> $facets[[4]]$edges
#>      [,1] [,2]
#> [1,]    2    3
#> [2,]    2    4
#> [3,]    3    5
#> [4,]    4    5
#> 
#> $facets[[4]]$ridges
#> [1]  1  5 10 11
#> 
#> $facets[[4]]$neighbors
#> [1] 1 2 5 6
#> 
#> $facets[[4]]$volume
#> [1] 1
#> 
#> $facets[[4]]$center
#> [1] 0.0 0.5 0.5
#> 
#> $facets[[4]]$normal
#> [1] -1  0  0
#> 
#> $facets[[4]]$offset
#> [1] 0
#> 
#> $facets[[4]]$family
#> [1] NA
#> 
#> $facets[[4]]$orientation
#> [1] 1
#> 
#> 
#> $facets[[5]]
#> $facets[[5]]$vertices
#> [1] 5 8 9 4
#> 
#> $facets[[5]]$edges
#>      [,1] [,2]
#> [1,]    4    5
#> [2,]    4    8
#> [3,]    5    9
#> [4,]    8    9
#> 
#> $facets[[5]]$ridges
#> [1]  3  8 11 12
#> 
#> $facets[[5]]$neighbors
#> [1] 1 3 4 6
#> 
#> $facets[[5]]$volume
#> [1] 1
#> 
#> $facets[[5]]$center
#> [1] 0.5 1.0 0.5
#> 
#> $facets[[5]]$normal
#> [1] 0 1 0
#> 
#> $facets[[5]]$offset
#> [1] -1
#> 
#> $facets[[5]]$family
#> [1] NA
#> 
#> $facets[[5]]$orientation
#> [1] -1
#> 
#> 
#> $facets[[6]]
#> $facets[[6]]$vertices
#> [1] 5 7 9 3
#> 
#> $facets[[6]]$edges
#>      [,1] [,2]
#> [1,]    3    5
#> [2,]    3    7
#> [3,]    5    9
#> [4,]    7    9
#> 
#> $facets[[6]]$ridges
#> [1]  6  9 10 12
#> 
#> $facets[[6]]$neighbors
#> [1] 2 3 4 5
#> 
#> $facets[[6]]$volume
#> [1] 1
#> 
#> $facets[[6]]$center
#> [1] 0.5 0.5 1.0
#> 
#> $facets[[6]]$normal
#> [1] 0 0 1
#> 
#> $facets[[6]]$offset
#> [1] -1
#> 
#> $facets[[6]]$family
#> [1] NA
#> 
#> $facets[[6]]$orientation
#> [1] 1
#> 
#> 
#> 
#> $volume
#> [1] 1
#> 
#> attr(,"3d")
#> [1] TRUE